requirements.txt
chromadb==0.5.11
langchain_community==0.3.1
langchain_core==0.3.9
langchain_groq==0.2.0
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.38.0
